# rest-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RESTful Template",
    "version": "1.0.0",
    "description": "RESTful Template API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/auth/logout": {
      "post": {
        "summary": "Logout user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "404": {
            "description": "Token not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/merchant/getMerchantAPI": {
      "post": {
        "summary": "Get merchant API details",
        "tags": [
          "Merchant"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string"
                  },
                  "userID": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merchant API details retrieved successfully"
          },
          "404": {
            "description": "Merchant not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/merchant/updateMerchantAPI": {
      "post": {
        "summary": "Update merchant API details",
        "tags": [
          "Merchant"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string"
                  },
                  "userID": {
                    "type": "string"
                  },
                  "marketplace-API": {
                    "type": "object",
                    "properties": {
                      "trendyolSupplierID": {
                        "type": "string"
                      },
                      "trendyolAPIKey": {
                        "type": "string"
                      },
                      "trendyolAPISecretKey": {
                        "type": "string"
                      },
                      "getirYemekMerchantToken": {
                        "type": "string"
                      },
                      "yemekSepetiUsername": {
                        "type": "string"
                      },
                      "yemekSepetiPassword": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marketplace API details updated successfully"
          },
          "404": {
            "description": "Merchant not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}